Fixed Delete Cita Past

+++ src/main/java/org/springframework/samples/petclinic/web/CitaController.java
@@ -36,7 +36,6 @@ public class CitaController {
 	private PacienteService pacienteService;
 	@Autowired
 	private UserService userService;
-	
 
 	// CUANDO ALGUIEN HAGA EL CITADETAILS POR FAVOR QUE USE ESTA URL EN EL MAPPING:
 	// "/citas/citaDetails/{citaId}"
@@ -51,17 +50,17 @@ public class CitaController {
 	public String listadoCitas(final ModelMap modelMap, @PathVariable("medicoId") final int medicoId) {
 		String vista = "citas/listCitas";
 		int idMedico = this.userService.getCurrentMedico().getId();
-		
-		if(medicoId != idMedico){
+
+		if (medicoId != idMedico) {
 			return "accessNotAuthorized";
 		} else {
 			Collection<Cita> citas = this.citaService.findCitasByMedicoId(medicoId);
-			if(citas.isEmpty()){
+			if (citas.isEmpty()) {
 				return "redirect:/";
 			} else {
-			modelMap.put("selections", citas);
-			modelMap.put("dateOfToday", LocalDate.now());
-			return vista;
+				modelMap.put("selections", citas);
+				modelMap.put("dateOfToday", LocalDate.now());
+				return vista;
 			}
 		}
 
@@ -74,11 +73,11 @@ public class CitaController {
 		Medico medic = this.userService.getCurrentMedico();
 		// cita.setPaciente(paciente);
 		// cita.setName("paciente");
-		
-		if(paciente.getMedico().equals(medic)){
-		modelMap.addAttribute("paciente", paciente);
-		modelMap.addAttribute("cita", cita);
-		return "citas/createOrUpdateCitaForm";
+
+		if (paciente.getMedico().equals(medic)) {
+			modelMap.addAttribute("paciente", paciente);
+			modelMap.addAttribute("cita", cita);
+			return "citas/createOrUpdateCitaForm";
 		} else {
 			return "accessNotAuthorized";
 		}
@@ -87,7 +86,7 @@ public class CitaController {
 	@PostMapping(path = "/save")
 	public String salvarCita(@Valid final Cita cita, final BindingResult result, final ModelMap modelMap)
 			throws InvalidAttributeValueException {
-			
+
 		if (result.hasErrors()) {
 			modelMap.addAttribute("cita", cita);
 			modelMap.addAttribute("paciente", cita.getPaciente());
@@ -98,7 +97,7 @@ public class CitaController {
 			modelMap.addAttribute("message", "La fecha debe estar en presente o futuro");
 			result.rejectValue("fecha", "error.fecha", "La fecha debe estar en presente o futuro");
 			return "citas/createOrUpdateCitaForm";
-		} else if (this.citaService.existsCitaPacienteDate(cita.getFecha(), cita.getPaciente())){
+		} else if (this.citaService.existsCitaPacienteDate(cita.getFecha(), cita.getPaciente())) {
 			modelMap.addAttribute("cita", cita);
 			modelMap.addAttribute("paciente", cita.getPaciente());
 			result.rejectValue("fecha", "error.citasamedate", "Ya existe una cita para este paciente en esta fecha");
@@ -114,19 +113,24 @@ public class CitaController {
 	public String borrarCita(@PathVariable("citaId") final int citaId, final ModelMap modelMap) {
 		Medico actualMedic = this.userService.getCurrentMedico();
 		Optional<Cita> cita = this.citaService.findCitaById(citaId);
-		
-		
-			if (cita.isPresent() && cita.get().getInforme() == null && actualMedic.equals(cita.get().getPaciente().getMedico())) {
-			modelMap.addAttribute("message", "Cita successfully deleted");
-			this.citaService.delete(cita.get());
-			return "redirect:/citas";
-			} else if(cita.isPresent() && !actualMedic.equals(cita.get().getPaciente().getMedico())){
-			return "accessNotAuthorized";
+
+		if (cita.isPresent() && cita.get().getInforme() == null
+				&& actualMedic.equals(cita.get().getPaciente().getMedico()) && cita.get().getFecha() != null) {
+			if (cita.get().getFecha().isAfter(LocalDate.now()) || cita.get().getFecha().isEqual(LocalDate.now())) {
+				modelMap.addAttribute("message", "Cita successfully deleted");
+				this.citaService.delete(cita.get());
+
+				return "redirect:/citas";
 			} else {
+				return "redirect:/citas";
+			}
+		} else if (cita.isPresent() && !actualMedic.equals(cita.get().getPaciente().getMedico())) {
+			return "accessNotAuthorized";
+		} else {
 			modelMap.addAttribute("message", "Cita cant be deleted");
 			return "redirect:/citas";
-			}
 		}
+	}
 
 	@GetMapping(value = "/find")
 	public String initFindForm(final Map<String, Object> model) {
